# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.Proxy:
    properties:
      allow_users:
        type: string
      auto_cert:
        description: 自动证书字段
        type: boolean
      bandwidth_limit:
        type: string
      bandwidth_limit_mode:
        type: string
      cert_id:
        description: 关联的证书ID
        type: integer
      client_id:
        type: integer
      created_at:
        type: string
      current_bytes_in_rate:
        type: integer
      current_bytes_out_rate:
        type: integer
      custom_domains:
        type: string
      dns_provider_id:
        description: DNS提供商ID
        type: integer
      dns_root_domain:
        description: 根域名
        type: string
      enable_dns_sync:
        description: DNS同步字段
        type: boolean
      enabled:
        type: boolean
      frp_cur_conns:
        type: integer
      frp_last_close_time:
        type: string
      frp_last_start_time:
        type: string
      frp_status:
        type: string
      health_check_interval:
        type: integer
      health_check_timeout:
        type: integer
      health_check_type:
        type: string
      host_header_rewrite:
        type: string
      http_password:
        type: string
      http_user:
        type: string
      id:
        type: integer
      last_online_time:
        type: string
      last_traffic_update:
        type: string
      local_ip:
        type: string
      local_port:
        type: integer
      locations:
        type: string
      name:
        type: string
      plugin_config:
        description: 插件配置 JSON
        type: string
      plugin_type:
        description: 插件配置字段
        type: string
      remote_port:
        type: integer
      secret_key:
        type: string
      subdomain:
        type: string
      total_bytes_in:
        type: integer
      total_bytes_out:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      use_compression:
        type: boolean
      use_encryption:
        type: boolean
    type: object
...
  util.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: FRP 管理面板后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FRP 管理面板 API
  version: "1.0"
paths:
...
  /api/clients/{id}/export:
    get:
      description: 导出指定客户端的FRP配置文件(frpc.toml格式)
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: FRP配置文件内容
          schema:
            type: string
        "500":
          description: 导出配置失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 导出客户端配置
      tags:
      - 代理管理
...
  /api/clients/{id}/proxies:
    get:
      consumes:
      - application/json
      description: 获取指定客户端下的所有代理配置
      parameters:
      - description: 客户端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 代理列表
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 获取代理列表失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取客户端代理列表
      tags:
      - 代理管理
...
  /api/proxies:
    get:
      consumes:
      - application/json
      description: 获取系统中所有代理配置的列表
      produces:
      - application/json
      responses:
        "200":
          description: 代理列表
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 获取代理列表失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取所有代理列表
      tags:
      - 代理管理
    post:
      consumes:
      - application/json
      description: 创建新的代理配置，客户端必须在线才能创建
      parameters:
      - description: 代理配置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Proxy'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误或客户端离线
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 创建代理
      tags:
      - 代理管理
  /api/proxies/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的代理配置，可选择是否同时删除关联的DNS记录，客户端必须在线才能删除
      parameters:
      - description: 代理ID
        in: path
        name: id
        required: true
        type: integer
      - default: true
        description: 是否同时删除DNS记录
        in: query
        name: deleteDNS
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: 客户端离线
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 删除代理失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 删除代理
      tags:
      - 代理管理
    put:
      consumes:
      - application/json
      description: 更新指定代理的配置，客户端必须在线才能更新
      parameters:
      - description: 代理ID
        in: path
        name: id
        required: true
        type: integer
      - description: 代理配置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Proxy'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误或客户端离线
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 更新代理失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 更新代理
      tags:
      - 代理管理
  /api/proxies/{id}/toggle:
    put:
      consumes:
      - application/json
      description: 切换指定代理的启用/禁用状态，客户端必须在线才能切换
      parameters:
      - description: 代理ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 切换成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 客户端离线
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 切换代理状态失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 切换代理启用/禁用状态
      tags:
      - 代理管理
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"